Move data parallel APIs
